debug: add log to track splits

+++ server/src/operators/file_operator.rs
@@ -119,7 +119,7 @@ pub async fn create_file_chunks(
             })
         })
         .transpose()?;
-    let target_splits_per_chunk = upload_file_data.target_splits_per_chunk.unwrap_or(20);
+    let target_splits_per_chunk = upload_file_data.target_splits_per_chunk.unwrap_or(20) as usize;
     let chunk_htmls = coarse_doc_chunker(file_text, split_regex, target_splits_per_chunk);
 
     let mut chunks: Vec<ChunkData> = [].to_vec();
